refactor: move dev tools class to dev flavor code
