+++ app/src/dev/java/xyz/lbres/trickcalculator/DeveloperToolsDialog.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.trickcalculator.ui.shared
+package xyz.lbres.trickcalculator
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -12,15 +12,16 @@ import android.widget.ArrayAdapter
 import android.widget.Spinner
 import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.ViewModelProvider
-import xyz.lbres.trickcalculator.MainActivity
-import xyz.lbres.trickcalculator.R
 import xyz.lbres.trickcalculator.databinding.DialogDeveloperToolsBinding
 import xyz.lbres.trickcalculator.ui.settings.initSettingsDialog
+import xyz.lbres.trickcalculator.ui.shared.SharedViewModel
 import xyz.lbres.trickcalculator.utils.gone
 import xyz.lbres.trickcalculator.utils.visible
 
+/**
+ * Dialog with various developer tools, only available in dev build flavor
+ */
 class DeveloperToolsDialog : DialogFragment() {
-    // TODO move to dev folder
     private lateinit var binding: DialogDeveloperToolsBinding
     private lateinit var viewModel: SharedViewModel
