+++ app/src/final/java/xyz/lbres/trickcalculator/BuildFlavor.kt
@@ -1,7 +1,9 @@
 package xyz.lbres.trickcalculator
 
+/**
+ * Configuration for final build flavor
+ */
 object BuildFlavor : BuildFlavorConfig {
     override val devMode = false
-
     override fun setupFlavor(activity: MainActivity) {}
 }
